OLS-UI handle too many index.js files
