+++ OLS-UI/ols-ui/src/components/AccountItem/AccountItem.js
@@ -1,0 +1,43 @@
+// libs
+import React from 'react';
+import classNames from 'classnames/bind';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faCheckCircle } from '@fortawesome/free-solid-svg-icons';
+import { Link } from 'react-router-dom';
+import PropTypes from 'prop-types';
+
+// components
+import styles from './AccountItem.module.scss';
+import Image from '~/components/Image';
+
+// account item là những Account items được xuất hiện trong list gợi ý của Popper
+/*
+- avatar 
+- name
+- tich xanh
+*/
+
+const cx = classNames.bind(styles);
+
+const AccountItem = ({ data }) => {
+    return (
+        <Link to={`/@${data.nickname}`} className={cx('wrapper')}>
+            <Image className={cx('avatar')} src={data.avatar} alt={data.full_name} />
+            <div className={cx('info')}>
+                <h4 className={cx('name')}>
+                    <span>{data.full_name}</span>
+
+                    {/* icon tich xanh */}
+                    {data.tick && <FontAwesomeIcon className={cx('check')} icon={faCheckCircle} />}
+                </h4>
+                <span className={cx('username')}>{data.nickname}</span>
+            </div>
+        </Link>
+    );
+};
+
+AccountItem.propTypes = {
+    data: PropTypes.object.isRequired,
+};
+
+export default AccountItem;
