+++ OLS-UI/ols-ui/src/layouts/DefaultLayout/index.js
@@ -1,29 +1,1 @@
-// libs
-import React from 'react';
-import classNames from 'classnames/bind';
-import PropTypes from 'prop-types';
-
-// components
-import styles from './DefaultLayout.module.scss';
-import Header from '~/layouts/components/Header';
-//import Sidebar from './Sidebar';
-
-const cx = classNames.bind(styles);
-
-const DefaultLayout = ({ children }) => {
-    return (
-        <div className={cx('wrapper')}>
-            <Header />
-            <div className={cx('container')}>
-                {/* <Sidebar /> */}
-                <div className={cx('content')}>{children}</div>
-            </div>
-        </div>
-    );
-};
-
-DefaultLayout.propTypes = {
-    children: PropTypes.node.isRequired,
-};
-
-export default DefaultLayout;
+export { default } from './DefaultLayout';
